BAVL-167: Moving response objects into the response folder and removing old validation rules.
