+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsbookavideolinkapi/service/mapping/PrisonAppointmentMappers.kt
@@ -1,6 +1,6 @@
 package uk.gov.justice.digital.hmpps.hmppsbookavideolinkapi.service.mapping
 
-import uk.gov.justice.digital.hmpps.hmppsbookavideolinkapi.model.PrisonAppointment
+import uk.gov.justice.digital.hmpps.hmppsbookavideolinkapi.model.response.PrisonAppointment
 import uk.gov.justice.digital.hmpps.hmppsbookavideolinkapi.entity.PrisonAppointment as PrisonAppointmentEntity
 
 fun PrisonAppointmentEntity.toModel() = PrisonAppointment(
